temporary fix the dropdown issue
the issue with the dropdown in this case is that the blank option presented here has a value of -1 while all other values are strings. this can be resolved in the example with the included change but I am continuing to look at the datagrid with the aim of understanding it better and providing a more durable solution to the problem.